grocery-delivery-platform/backend/consumer-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
pydantic[email]==1.8.2
python-jose[cryptography]==3.3.0
python-dotenv==0.19.0
httpx==0.19.0
elasticsearch==7.15.0 